moving api crud operations into their own file
